[Lumen Laravel] Buil docker

+++ routes/web.php
@@ -20,44 +20,27 @@ $router->get('/', function () use ($router) {
     return $router->app->version();
 });
 
-$router->post('/login', 'AuthController@login');
-$router->post('/logout', ['middleware' => 'auth', 'uses' => 'AuthController@logout']);
+$router->group(['middleware' => 'cors'], function () use ($router) {
+    $router->post('/login', 'AuthController@login');
+    $router->post('/logout', ['middleware' => 'auth', 'uses' => 'AuthController@logout']);
 
-// Add auth middleware to all routes
-$router->group(['middleware' => 'auth:api'], function () use ($router) {
-    // Admin routes
-    $router->group(['prefix' => 'admin'], function () use ($router) {
-        $router->get('/customers', [
-            'middleware' => 'role:0',
-            'uses' => 'CustomerController@index'
-        ]);
-        $router->post('/customers', [
-            'middleware' => 'role:0',
-            'uses' => 'CustomerController@store'
-        ]);
-        $router->put('/customers/{id}', [
-            'middleware' => 'role:0',
-            'uses' => 'CustomerController@update'
-        ]);
-        $router->delete('/customers/{id}', [
-            'middleware' => 'role:0',
-            'uses' => 'CustomerController@destroy'
-        ]);
-    });
+    $router->group(['middleware' => 'auth:api'], function () use ($router) {
+        // Admin routes
+        $router->group(['prefix' => 'admin'], function () use ($router) {
+            $router->get('/customers', ['middleware' => 'role:0', 'uses' => 'CustomerController@index']);
+            $router->get('/orders', ['middleware' => 'role:0', 'uses' => 'CustomerController@allOrders']);
+            $router->post('/customers', ['middleware' => 'role:0', 'uses' => 'CustomerController@store']);
+            $router->put('/customers/{id}', ['middleware' => 'role:0', 'uses' => 'CustomerController@update']);
+            $router->put('/change-password', ['middleware' => 'role:0', 'uses' => 'CustomerController@changePassword']);
+            $router->delete('/customers/{id}', ['middleware' => 'role:0', 'uses' => 'CustomerController@destroy']);
+        });
 
-    // Customer routes
-    $router->group(['prefix' => 'customer'], function () use ($router) {
-        $router->get('/', [
-            'middleware' => 'role:1',
-            'uses' => 'CustomerController@show'
-        ]);
-        $router->put('/', [
-            'middleware' => 'role:1',
-            'uses' => 'CustomerController@updateProfile'
-        ]);
-        $router->get('/orders', [
-            'middleware' => 'role:1',
-            'uses' => 'CustomerController@orders'
-        ]);
+        // Customer routes
+        $router->group(['prefix' => 'customer'], function () use ($router) {
+            $router->get('/info', ['middleware' => 'role:1', 'uses' => 'CustomerController@show']);
+            $router->put('/', ['middleware' => 'role:1', 'uses' => 'CustomerController@updateProfile']);
+            $router->put('/change-password', ['middleware' => 'role:1', 'uses' => 'CustomerController@changePasswordCustomer']);
+            $router->get('/orders', ['middleware' => 'role:1', 'uses' => 'CustomerController@orders']);
+        });
     });
 });
